Update Directory Structure and Add Play Functionality
