+++ lib/core/error/failures.dart
@@ -1,0 +1,10 @@
+import 'package:equatable/equatable.dart';
+
+abstract class Failure extends Equatable {
+  final List properties;
+
+  Failure([this.properties = const <dynamic>[]]);
+
+  @override
+  List<Object> get props => [properties];
+}
