+++ lib/core/usecases/usecases.dart
@@ -1,0 +1,6 @@
+import 'package:dartz/dartz.dart';
+import 'package:earbender/core/error/failures.dart';
+
+abstract class UseCase<Type, Params> {
+  Future<Either<Failure, Type>> call(Params params);
+}
