+++ lib/core/utils/no_glow_scholl_behavior.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+class NoGlowScrollBehavior extends ScrollBehavior {
+  @override
+  Widget buildViewportChrome(
+      BuildContext context, Widget child, AxisDirection axisDirection) {
+    return child;
+  }
+}
