+++ lib/features/main/presentation/widgets/background_color_grid_tile_widget.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class ColoredTiles extends StatelessWidget {
+  final Color tileColor;
+  ColoredTiles({this.tileColor});
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      color: tileColor,
+    );
+  }
+}
